requirements.txt
flask==3.0.3
python-dotenv==1.0.1
langchain
langchain-openai
google-genai
